requirements.txt
backports.tarfile==1.2.0
elasticsearch==8.17.2
google-cloud-storage==3.1.0
google-genai==1.9.0
importlib-metadata==8.0.0
ipykernel==6.29.5
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.5
jaraco.collections==5.1.0
pandas==2.2.3
pathspec==0.12.1
pdf2image==1.17.0
pip-chill==1.0.3
pipreqs==0.5.0
pycryptodome==3.22.0
pymysql==1.1.1
pypdf2==3.0.1
python-dotenv==1.1.0
sqlalchemy==2.0.40
tinycss2==1.4.0
tomli==2.0.1
tqdm==4.67.1
